+++ cmd/manager/Dockerfile
@@ -22,9 +22,9 @@ COPY go.sum go.sum
 RUN go mod download
 
 COPY cmd/manager/main.go cmd/manager/main.go
-COPY api api/
-COPY controllers controllers/
-COPY pkg pkg/
+COPY api/ api/
+COPY internal/controller internal/controller
+COPY pkg/ pkg/
 
 RUN CGO_ENABLED=0 GOOS=${TARGETOS:-linux} GOARCH=${TARGETARCH} go build -a -o manager cmd/manager/main.go
